requirements.txt
# Google ADK is the heart of the multi-agent system
google-adk==0.4.0

# Gemini model access via Google's AI SDK
google-generativeai==0.7.2

# FastAPI for the API layer (ADK can also serve directly, but we'll wrap it)
fastapi==0.111.0
uvicorn==0.30.0

# Database layer — unchanged
sqlalchemy==2.0.30

# Utilities
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
pandas==2.2.2